Created smoketest verifying the bot starts
