+++ .travis.yml
@@ -3,3 +3,4 @@ language: node_js
 node_js:
   - 4
   - 5
+  - 6
