+++ lib/github-client.js
@@ -14,7 +14,7 @@ const githubClient = new GitHub({
 
 githubClient.authenticate({
   type: 'oauth',
-  token: process.env.GITHUB_TOKEN
+  token: process.env.GITHUB_TOKEN || 'invalid-placeholder-token'
 })
 
 module.exports = githubClient
