+++ lib/github-events.js
@@ -28,25 +28,10 @@ module.exports = (app) => {
 
     res.end()
 
-    emit(data)
+    app.emitGhEvent(data)
   })
 
-  if (process.env.SSE_RELAY) {
-    const EventSource = require('eventsource')
-    var es = new EventSource(process.env.SSE_RELAY)
-    es.onmessage = (e) => {
-      try {
-        const data = JSON.parse(e.data)
-        if (!data.action) return
-
-        emit(data)
-      } catch (e) {
-        console.error(e)
-      }
-    }
-  }
-
-  function emit (data) {
+  app.emitGhEvent = function emitGhEvent (data) {
     const repo = data.repository
     const org = repo.owner.login || data.organization.login
