+++ scripts/ping.js
@@ -1,0 +1,7 @@
+'use strict'
+
+module.exports = function (app) {
+  app.get('/ping', (req, res) => {
+    res.end('pong')
+  })
+}
